Added id checking on deleting a record

+++ routes/medical-records.route.js
@@ -77,6 +77,10 @@ router.route("/list").get((req, res) => {
 
 // Delete Medical Record
 router.route("/delete/:id").delete((req, res, next) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ msg: "Specified id is not valid" });
+  }
+
   recordSchema.findByIdAndRemove(req.params.id, (error, data) => {
     if (error) {
       return next(error);
